chore(clippy): fix pedantic lint (#1375)
* chore: unnecessary wraps


* chore: ignored unit patterns


* chore: needless for each


* chore: explicit iter loop


* chore: cast possible truncation


* chore: redundant else


* chore: must use candidate


* chore: struct excessive bools


* chore: needless continue


* chore: ref option


* chore: match wildcard for single variants


* chore: needless pass by value


* chore: format push string


* chore: single match else


* chore: non std lazy statics


* chore: needless raw string hashes


* revert: "chore: needless for each"

This reverts commit abdc8ac6c015c172cf79cadbae5522ff7cad8042.


* update: ci.yml


* revert: "chore: cast possible truncation"

This reverts commit 5a94e71fc2ef55c56ec903b266361d5842f60d3d.


* fix: MAX_PAGE_SIZE type as i32


* revert: "chore: redundant else"

This reverts commit b6a1b30f12084d18e4eb211224ddc68eb98cc5da.


* update: ci.yml


* fix: MAX_PAGE_SIZE type as i32


* revert: "chore: format push string"

This reverts commit 620a053c8a07619e3549228d890f898b57ea6a41.


* update: ci.yml


* update: PULL_REQUEST_TEMPLATE.md


* update: CONTRIBUTING.md


---------

+++ git-cliff-core/src/remote/mod.rs
@@ -49,7 +49,7 @@ pub(crate) const REQUEST_TIMEOUT: u64 = 30;
 pub(crate) const REQUEST_KEEP_ALIVE: u64 = 60;
 
 /// Maximum number of entries to fetch in a single page.
-pub(crate) const MAX_PAGE_SIZE: usize = 100;
+pub(crate) const MAX_PAGE_SIZE: i32 = 100;
 
 /// Trait for handling remote commits.
 pub trait RemoteCommit: DynClone {
